[security] Negotiate authentication type during RPC setup
This commit introduces the concept of an authentication type to the RPC
negotiation sequence. The current valid authentication types are SASL,
CERTIFICATE, and TOKEN. Early in the negotiation sequence the client and
server decide on an authentication type to use for the connection based
on the mutually supported capabilities. If either side does not support
the new authentication negotiation, the connection automatically falls
back to SASL authentication in order to maintain backwards
compatibility.

This commit also adds TSKs to the master<->tserver heartbeat protocol,
so that tokens may be verified on the server. The client's authn token
has been moved to the messenger to make it more accessible to
negotiation.

negotiation-test has been updated with a general test runner for
different negotiation configurations.  There are so many possible
negotiation configurations that an exhaustive set of tests is not
attempted, but it should be straightforward to add additional tests in
the future.

Change-Id: I8ed9a1a474990dbfe9b71173adffdec95ec02b6c
Reviewed-on: http://gerrit.cloudera.org:8080/5988
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins