[java client] Limit the message size in Status
We've seen cases where RPC traces that are getting logged are MBs in size.
Even if we limit the number of traces, the messages themselves can get quite big.

This fix is only for the Java client since it's the only client with tracing at
the moment. Ideally we'd limit this also in the server but it's still desirable
to have the clients protect themselves.

With the limits in place, we can have a maximum of 100 * 256 = 25KB of messages
in a single RPC's trace.

Change-Id: I5868da9ee6b4befd1dc863570b8e251369213ffa
Reviewed-on: http://gerrit.cloudera.org:8080/5956
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>