rpc: use TLS only for authentication for loopback connections
This adds a new feature flag TLS_AUTHENTICATION_ONLY which is advertised
by the client and server in the case that they detect that the remote
peer is on the same socket address as the local one. When both peers
advertise this flag, TLS authentication proceeds as usual, but after a
successful handshake the connection is _not_ wrapped with TLS.

In case any users take fault with this optimization, an escape hatch
flag is provided which can disable it.

This patch only implements this feature on the C++ side. The Java side
will be done in a future patch.

Change-Id: I5bb637b3b06de4c4571bf2411aa0aaf8e1f41c2f
Reviewed-on: http://gerrit.cloudera.org:8080/5976
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>