security: authorize all RPCs against coarse-grained ACLs
This adds two new flags: 'superuser_acl' and 'user_acl'.
Cluster-admin operations (eg things like SetFlags) are authorized
against superuser_acl, and read/write/DDL type operations are authorized
against user_acl.

Internal-facing RPCs are authorized against the service user, which is
assumed to be a matching principal across all of the hosts.

Most of the "service" RPCs are also allowed to be accessed by
superusers, so that operator tools can take advantage of them. The one
exception is TSHeartbeat, which is locked down to _only_ the service
user, since it's the endpoint that exports signed IPKI certs.

The default service user and superuser ACL are set based on the identity
of the server: if the server is logged in from a keytab, we use that
username. Otherwise, we use the local Unix username. This means that
tests which rely on superuser things like SetFlags, etc, should continue
to work as before, since the test client runs as the same Unix user as
the server.

A new unit test smoke tests the various authorization levels using a
combination of the real client and hand-crafted RPCs.

Change-Id: Id24a6429273aff355e70e127086a26b7e4a03cd8
Reviewed-on: http://gerrit.cloudera.org:8080/5998
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>