java: redo locking and simplify TabletClient
This gets rid of some of the attempts at fancy synchronization and
instead uses relatively more coarse-grained locking to protect the
important state fields of TabletClient.

This also gets rid of the special handling for different types of
exceptions which tried to avoid re-fetching tablet locations from the
master. This detection was apparently racey anyway, as the
channelDisconnected() event would often arrive before the
uncaughtException() event, which meant that we'd try to reconnect to the
same server over and over even if it was down.

Before this fix, bumping the debug level log messages in TabletClient
down to trace caused several tests to fail due to exposing the above
race more frequently. With the patch, the log levels can be changed
safely.

Change-Id: I6cc03572821dc3401a7c06108dd7376a3f3cff18
Reviewed-on: http://gerrit.cloudera.org:8080/6017
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins