KUDU-1880 Prevent eager ignoring of NULLs
During a nullable block scan, if a NULL set of cells is reached before
the decoder evaluation is disabled (e.g. if the first rows in the
block are NULL), the rows will be prematurely marked as not in the
results set. If evaluating an IsNull predicate, this will yield
erroneous results.

This patch adds an API to the materialization context that indicates
whether it is evaluating an IsNull predicate. If it is, the null rows
must not be removed from the results set. Otherwise, null rows can be
cleared immediately.

Additional tablet unit tests are added to cover the IsNull case with
NULL values at the beginning of columns for various types. These
tests will fail without the fixes in the patch and pass with them.

Change-Id: Ib418a4fcc2794ce2f686e864f51834fb4fb8b048
Reviewed-on: http://gerrit.cloudera.org:8080/6029
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins