KUDU-1844: /varz should not expose potentially sensitive configs
This adds a new flag tag 'sensitive' and a new flag --redact_sensitive_flags.

If a flag is tagged as 'sensitive', and --redact_sensitive_flags is set to
true, then its value would be redacted as '<redacted>' when using /varz to
dump the flags. So that it would not be exposed via WebUI.

Example output with flags redacted:
  --test_sensitive_flag=<redacted>

Change-Id: I7661e0409c79559f22b07e7c12737e804544186b
Reviewed-on: http://gerrit.cloudera.org:8080/6043
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>