java: support token authentication in negotiation
This adds support for the TOKEN authentication type in negotiation, as
well as a place to store the token and CA cert within the client. This
isn't quite hooked up to the ConnectToCluster RPC yet, so it won't
actually take effect for real connections, but a new unit test verifies
the code paths.

A later patch will actually hook this up to the real certs and tokens
once the appropriate server-side code has all landed.

Change-Id: I7da5809495d67b308828e83c857249b852f21db3
Reviewed-on: http://gerrit.cloudera.org:8080/6047
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>