java: further cleanup of ConnectToCluster
This changes the individual RPCs which go to masters to directly return
the ConnectToMasterResponsePB. Then, these responses are aggregated in
the ConnectToCluster class, and once a leader is found, a new
ConnectToClusterResponse object is created which represents the result
of the aggregated multi-master operation.

The code to translate from the ConnectToMasterResponsePB is also moved
into the ConnectToClusterResponse class where it's a little more obvious
what it's doing, and called from AsyncKuduClient.

This also addresses a bunch of references in comments to the old RPC
names which got missed in the IDE-driven rename.

Change-Id: Ia51175ee1b4108341959d90545e2d6733d5e80d5
Reviewed-on: http://gerrit.cloudera.org:8080/6053
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>