java: fetch and adopt authn token and CA from master
This makes the Java client fetch a token and CA information from the
master upon first connect, and also exposes APIs to export and import
the security information in a serialized protobuf form.

A new test connects to a Kerberized cluster and serializes this
authentication info. It then drops Kerberos credentials, and passes the
credentials to a new client, which is still able to connect to the
cluster using the serialized info.

The new APIs introduced here are marked Unstable since we may need to
get more experience using them in execution frameworks before locking
them down.

Change-Id: I24e71f0f5feefa36106a50074d40731911eff64b
Reviewed-on: http://gerrit.cloudera.org:8080/6077
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>