java: add a fake DNS resolver to avoid reverse DNS hangs
On my system, I found that reverse DNS lookups for IPs like 127.3.4.5
were causing 5-second hangs. This wasn't a problem previously because we
don't ourselves do any reverse DNS, but it turns out that the GSSAPI
implementation in Java always does reverse DNS with no way to disable
it.

This adds a wrapper which can be installed into the
InetAddress.nameServices list which short-circuits the reverse lookups
for such IP addresses. This fixed the 5-second negotiation hangs that I
saw on my system.

Change-Id: Ife71d7a815c7ef2028e569476fba060481b00f97
Reviewed-on: http://gerrit.cloudera.org:8080/6076
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>