client: support exporting/importing authentication data
This adds the C++ side of being able to export and import tokens. This
will be useful for Impala in particular -- the front-end can acquire a
token and pass it as part of the plan fragment descriptor to the
backends, so that the backends don't need to use Kerberos to
authenticate.

Change-Id: I846c47feab9a7be0dce625079193146bfc2dac7b
Reviewed-on: http://gerrit.cloudera.org:8080/6086
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>