thirdparty: upgrade to python 2.7.13
On very rare occasion, the Python build fails with an error like this:

  Parser/acceler.o: file not recognized: File truncated
  collect2: error: ld returned 1 exit status

It's not reproducible, and it seems to only happen when make is run with a
strange -j and the machine placed under load. The failure seems like
this [1] Python issue, a fix for which was backported to 2.7.12. So let's
take the opportunity to upgrade to the latest Python 2.7 release.

I tested this by rebuilding the thirdparty tree on a CentOS 6.6 box.

1. https://bugs.python.org/issue22359

Change-Id: I309bc6a0ef7d919d3610ed322d76953dfb090426
Reviewed-on: http://gerrit.cloudera.org:8080/6093
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>