rpc: add a new RemoteUser class for the server side user info
We were previously using UserCredentials both as a client-side and
server-side class to represent a user. However, we'd like to capture
more information on the server side (such as both the full principal as
well as the shortened "username", and maybe the authentication method,
etc).

This patch splits out a new RemoteUser class which is used for
server-side connections. I also simplified the existing UserCredentials
class a bit while I was at it.

Change-Id: I899dc3b63d1f68ebf4120f1f8f6ab4b6e21c1a18
Reviewed-on: http://gerrit.cloudera.org:8080/6105
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>