rpc: separately capture full principal name and short username
This changes GSSAPI authentication to capture the full principal in one
field of RemoteUser while storing the "short" username in the 'username'
field. This is so that admins can configure ACLs (and later group
mapping) based on the expected short usernames.

The "short user" mapping respects the auth_to_local rules configured in
krb5.conf. If no appropriate rule can be found, then we fall back to
Hadoop's default of just using the first part of the principal.

This also changes CERTIFICATE authentication to expose the new X509
extensions in the same corresponding fields of RemoteUser.

Additionally, the authentication method is preserved in the RemoteUser
object, which will allow for some special-case authorization decisions
(e.g. we shouldn't issue tokens to existing users of tokens).

Change-Id: Iaf76abf8348dd677a47330b6d0ef110129d853c0
Reviewed-on: http://gerrit.cloudera.org:8080/6106
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>