[linked list test] use timeout for WaitAndVerify in all modes
Constrain LinkedListTester::WaitAndVerify by the specified timeout
for FINISH_WITH_SNAPSHOT_SCAN mode as well.  Otherwise the
TestLoadAndVerify scenario of the LinkedListTest test can
run indefinitely under some conditions.  One of such conditions is
the expiration of the client authn token during the test: if that
happens, WaitAndVerify() call might end up doing indefinite retries.

The proper client behavior on the expiration of the authn token will be
addressed in a separate changelist.

Change-Id: If129738fbeef5394dfd1bcb160121e1d4131b3cc
Reviewed-on: http://gerrit.cloudera.org:8080/6155
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>