Fix LTO support
This fixes the CMake setup to be able to do LTO builds using lld and clang. I
was able to successfully enable LTO with this. However, the performance of
clang-3.9 + ThinLTO appeared to still be worse than devtoolset's gcc-4.9.2.

Nevertheless, it seemed like a good thing to commit to replace the currently
bogus flags.

Change-Id: I58a532dec43bd9ed5d7ea6419bc9dfc011f279b6
Reviewed-on: http://gerrit.cloudera.org:8080/6162
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>