KUDU-1899. Fix support for empty string keys
We've never prevented users from inserting a row with "" as a key. However, a
faulty assertion in diskrowset.cc caused the tablet server to crash on flush in
this case.

This adds a new end-to-end test via the client which inserts, updates, and deletes
a row with a "" key. It caused a crash prior to this patch.

Change-Id: I779f25afe6d39d91067b1e7c1238797ec2ac0295
Reviewed-on: http://gerrit.cloudera.org:8080/6163
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>