Fix TestKerberosRenewal/Reacquire flakiness
It has been observed that an ExternalMiniCluster can take quite
long to start up (max observerd: 6s). TestKerberosRenewal/Reacquire ran
with the kerberos ticket lifetime as 5 seconds. If the ExternalMiniCluster
took longer than 5 seconds to start, the ticket of the test client would
expire, causing an unexpected test failure.

This patch increases the ticket lifetime to 15s. This will increase the
test time, but reduces the chance for flakiness.

Change-Id: If03a2b47e78a22736311e8dd160ddf63b86c4396
Reviewed-on: http://gerrit.cloudera.org:8080/6168
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>