KUDU-1901. Fix crash in concurrent OpenSSL usage
This fixes a crash in which a TLSContext's SSL_CTX was being used by a
thread to accept a connection at the same time as its underlying
certificate was being changed.

A new unit test reproduced the error pretty reliably after looping for
10-20 seconds. With the new locking, it no longer produces.

I also built OpenSSL with TSAN locally and verified that, after the
change, the only races reported within OpenSSL were those where the code
mentions the possibility of benign races.

Change-Id: I578350ba6a492e6e3ef635e9294f25fc0dc9d125
Reviewed-on: http://gerrit.cloudera.org:8080/6187
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>