KUDU-1904 Don't seek on empty RLE blocks
Seeking in RLE blocks enforces that the seek position is less than
the number of elements in the block. If the number of elements is 0,
as it is when the entire set of cells is NULL, this check will always
result in a crash.

This patch enforces that the block only seeks if there are elements
in the block.

Test coverage will be added in an upcoming commit.

Change-Id: I6d6bd95099c9f30aa923a0da1e76f92ed90ebd99
Reviewed-on: http://gerrit.cloudera.org:8080/6184
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins