[delete_table-test] Don't fail on multiple attempts to delete an orphaned tablet
TestUnknownTabletsAreNotDeleted is flaky with the following assertion:
/data/jenkins-workspace/kudu-workspace/src/kudu/integration-tests/delete_table-test.cc:1099: Failure
Value of: num_delete_attempts
  Actual: 3
Expected: 1

This is due to the master receiving more than 1 tablet report before
the tablet server has the time to delete the tablet. Security-related
changes likely shifted the timings and the heartbeat period is set
very low in this test, so this is more likely to happen. Nonetheless
the opportunity for flakyness was there even before these changes.

In addition to fixing the assertion on the metric, this patch also
adds an assertion that the tablet was actually deleted.

Change-Id: I2b5f21c718bad8c0b7112395082371f19a61507f
Reviewed-on: http://gerrit.cloudera.org:8080/6191
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: David Ribeiro Alves <dralves@apache.org>