[delete_table-test] Disable log compression on some tests
Some tests rely on log rolls but with compression enabled these
might take a while. So long that it might trip test timeouts.
I inspected the logs and the test failures this addresses don't
have another visible cause. The log is silent for a full minute
while writes are going on and then the wait fails on a timeout.

This just disables compression in these cases.

Change-Id: I8630cc73f431f5f2e4f567ae02733cdafe073666
Reviewed-on: http://gerrit.cloudera.org:8080/6200
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins