Make ExternalDaemon::StartProcess() handle fault injection
Occasionally starting a kudu daemon fails becuase of fault injection.
We recently fixed this error manifesting as SIGPIPE, but we didn't
fix the underlying cause: under load, tests doing fault
injection might fail before StartProcess() completes making
this method return a non-OK() status.

This patch makes sure that if the exit status is an expected one (85)
we still return as if starting the process had succeeded. Subsequent
calls to WaitForInjectedCrash() will return Status::OK();

To make sure this works, this patch includes a test that enables
a "fake" fault early in the startup process and makes sure that
both WaitForInjectedCrash and StartProcess() return OK() in
spite of it.

Change-Id: I6046e34a321de3e324e20e3d63249e4073712447
Reviewed-on: http://gerrit.cloudera.org:8080/6211
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins