[ts_recovery-itest] Reduce flakyness of TestCrashBeforeWriteLogSegmentHeader
This test expects a log roll before crashing, but with log compression
enabled we can actually write a lot of data before rolling. This is
making the test fail later with a timeout on Restart().

This patch disables log compression for this test, making sure that
we don't write as much and thus making restart faster.

Change-Id: I208fc1fb05dcb0cd6f5386792efc66e5e24272de
Reviewed-on: http://gerrit.cloudera.org:8080/6279
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins