[consensus] add lock coverage for peer queue
A Peer or PeerQueue would lock without taking into account the
presence of a peer calling back after TabletCopy and updating its
members, which doesn't abide by external locks (like the queue lock).
This would result in races.

A PeerMessageQueue operation would lock with the assumption that its
access to a peer was uninterrupted by other threads. This assumption
is broken in cases where a PeerProxy object has a callback to access a
peer (e.g. proxy calls StartTabletCopy with a callback that processes
the response and updates the peer) from a separate thread (as is the
case with the RpcHeartBeater).

A Peer operation like SendNextRequest would lock under this assumption,
only to potentially be interrupted by the same callback. This callback
was not completely locked and would race with the send.

This patch adds lock coverage to these cases, both seen in
RaftConsensusITest.TestConfigChangeUnderLoad.

TSAN errors and descriptions can be found here:
https://github.com/andrwng/kudu/blob/consensus-locks/docs/race_diagnostics/race_descriptions.txt

Full TSAN logs can be found here:
https://github.com/andrwng/kudu/blob/consensus-locks/docs/race_diagnostics/race_type_1.txt
https://github.com/andrwng/kudu/blob/consensus-locks/docs/race_diagnostics/race_type_2.txt

Dist-test results before and after:
Before: Note that for these builds, status 1 is likely a data race.
http://dist-test.cloudera.org/job?job_id=awong.1489029805.26963
http://dist-test.cloudera.org/job?job_id=awong.1489086157.18460

After:
http://dist-test.cloudera.org/job?job_id=awong.1489096815.24680

Change-Id: I0e3f9603c471441d9b926cac3377fc6a5a2a1514
Reviewed-on: http://gerrit.cloudera.org:8080/6299
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins