[flaky tests] Address LSAN false positives
Some tests, like external_mini_cluster-itest, are now only failing
in ASAN due to leaks. Upon inspection I could only find false
positives among these leaks around thread local variables.

This seems related to this issue:
https://github.com/google/sanitizers/issues/757

The fix is to wrap those specific instances of false positives
with ScopedLeakDisabler so that they don't get reported.

After many tries I was unable to come up with the right incantation
to cause this in dist-tests. Running the highest failing test in
ASAN with stress was not enough to cause this. Somehow this is
more likely in jenkins where these failures are pretty common.

Change-Id: I8c8d9ff83c0cfbc11cab213a25cbd5daa3b25869
Reviewed-on: http://gerrit.cloudera.org:8080/6320
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>
Tested-by: Kudu Jenkins